services/policy-service/requirements.txt
aiohttp==3.9.1
pyyaml==6.0.1
